update installed packages view
